mor/requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
wandb>=0.15.0
pyyaml>=6.0
numpy>=1.24.0
tqdm>=4.65.0
einops>=0.6.0
matplotlib>=3.7.0
seaborn>=0.12.0
rich>=13.0.0
pytest>=7.3.0
